# docs/swagger.yaml
openapi: 3.0.0
info:
  title: Spotplay
  description: 'Es una plataforma de musica y mas ...'
  contact:
    email: suport@spotplay.com
  license:
    name: Apache 2.0
    url: http://apac.....
  version: 1.0.0
servers:
- url: http://localhost:4000/api/v1
...
components:
...
  schemas:
...
    User:
      type: object
      properties:
        id:
          type: integer
          format: int64
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
...
    UserPost:
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        role:
          type: string
    UserDelete: 
      type: object
      properties:
        id: 
          type: string
...
paths:
...
  /user/signup:
    post:
      tags:
      - User
      summary: Create user
      operantionId: createUser
      description: New user is created, sends Welcome message to the user email. 
      requestBody:
        description: ingrese los datos del usuario a crear, si es usuario lector Role:lector
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserPost'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserPost'
        required: true
      responses:
        200:
          description: registro exitoso!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPost'
        400:
          description: bad request
          content: {}
        404:
          description: not found
          content: {}
  /user/:
    get:
      tags:
      - User
      summary: get all Users
      operantionId: getAllUsers
      description: See all users
      responses:
        200:
          description: registro exitoso!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserPost'
        400:
          description: bad request
          content: {}
        404:
          description: not found
          content: {}
    delete:
      tags:
      - User
      summary: Delete current user
      operantionId: deleteUser
      description: Delete user
      requestBody:
        description: Write all data of user to delete
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserDelete'
          application/xml:
            schema:
              $ref: '#/components/schemas/UserDelete'
        required: true
      responses:
        200:
          description: registro exitoso!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: bad request
          content: {}
        404:
          description: not found
          content: {}
    put:
      tags:
      - User
      summary: Update User
      operantionId: updateUser
      description: update a given user.
      requestBody:
        description: write the information of user to update.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
          application/xml:
            schema:
              $ref: '#/components/schemas/User'
        required: true
      responses:
        200:
          description: registro exitoso!
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        400:
          description: bad request
          content: {}
        404:
          description: not found
          content: {}
...